remove a misleading comment
The whole intention within action's input/output handling here was to
put some emphasis on the fact that we aren't tracking the input/output
nodes directly. However, looking at all the other classes this is a bit
redundant, we're using a pattern like this all the time. Just drop the
comment.

Change-Id: Ibd9bf9f1e83c650dda3bc43ef48e61dd6d95da5a