disable building tests for hh-date on new enough GCC
On GCC 14.2, this won't build for me:

 include/libyang-cpp/Time.hpp:117:34: error: no matching function for call to ‘parse(const char [21], std::chrono::time_point<std::chrono::_V2::system_clock, std::chrono::duration<long int, std::ratio<1, 1000000000> > >&)’

The error message is actually the same as in a proposed fix [1], and
upstream maintainer has said that they actually recommend relying on the
STL on this "new enough" platform. I'm very happy to oblige.

We were discussing this last year
(I51242fe2ec873fba5954c44e38e898d205a9e484), but at the time this was
implemented the build failure was not known, AFAICT, so let's assume
that the former use case for building with "new enough STL", but
actually forcing HH-date, is no longer important.

[1] https://github.com/HowardHinnant/date/pull/866

Change-Id: Id4b0206e2a960686fb1d8e76134b9061d7545d81