port to libyang v4
Apart from the usual API/ABI changes, the default values of a `leaf` or
a `leaf-list` are no longer canonicalized.

This used to be missing from libyang v4, but it was added recently.
Sill, chances are that the form which was used in the YANG model's
source code is actually *the* intended presentation, so let's skip
canonicalization altogether.

Upstream has removed some flags and defines, so let's sync their list
with whatever is currently available in the `devel` branch.

Change-Id: I52dcec88eeee003b45dd6cf400f4d6875abbcc05